feat: ethers instantiation with Reown
